Updated argument list of ruCubeSingleSolveHandlerPool. buffer size and number of cubes to fetch at a time can be set

+++ src/ruCubeMultiSolveHandler.h
@@ -80,7 +80,7 @@ class ruCubeMultiSolveHandler
 
 
         uint32_t totalNumberOfCubesToGenerate = 0;
-        uint8_t numOfThreads = 1;
+        size_t numOfThreads = 1;
 };
 
 #endif // RUCUBEMULTISOLVEHANDLER_H
